% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_measure_update.R
\name{BaseMeasureUpdate}
\alias{BaseMeasureUpdate}
\title{Update the parameters of the base measure ($G_0$) using
the current cluster parameter samples.}
\usage{
BaseMeasureUpdate(dpobj)
}
\arguments{
\item{dpobj}{Dirichlet process object}
}
\description{
Update the parameters of the base measure ($G_0$) using
the current cluster parameter samples.
}
