% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testHelpers.R
\name{benchmark_r_cpp}
\alias{benchmark_r_cpp}
\title{Benchmark R vs C++ implementation}
\usage{
benchmark_r_cpp(r_func, cpp_func, ..., times = 100)
}
\arguments{
\item{r_func}{R function to benchmark}

\item{cpp_func}{C++ function to benchmark}

\item{...}{Arguments to pass to both functions}

\item{times}{Number of repetitions}
}
\value{
Benchmark results
}
\description{
Benchmark R vs C++ implementation
}
\keyword{internal}
