% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_uniform_pareto.R
\name{BetaMixture2Create}
\alias{BetaMixture2Create}
\title{Create a Beta mixture with zeros at the boundaries.}
\usage{
BetaMixture2Create(priorParameters = 2, mhStepSize = c(1, 1), maxT = 1)
}
\arguments{
\item{priorParameters}{The prior parameters for the base measure.}

\item{mhStepSize}{The Metropolis Hastings step size. A numeric vector of length 2.}

\item{maxT}{The upper bound of the Beta distribution. Defaults to 1 for the standard Beta distribution.}
}
\value{
A mixing distribution object.
}
\description{
This bounded Beta mixing distribution also uses the mean and precision
parameterisation on \eqn{(0, maxT)}.
}
