% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R, R/fit.R
\name{can_use_cpp}
\alias{can_use_cpp}
\title{Check if C++ implementation is available for this model}
\usage{
can_use_cpp(dp_obj)

can_use_cpp(dp_obj)
}
\description{
Check if C++ implementation is available for this model

Check if C++ implementation is available for this model
}
\keyword{internal}
