% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_cpp_interface.R
\name{can_use_hierarchical_cpp}
\alias{can_use_hierarchical_cpp}
\title{Check if hierarchical C++ implementation is available}
\usage{
can_use_hierarchical_cpp(dp_list)
}
\arguments{
\item{dp_list}{Hierarchical DP object}
}
\value{
Logical indicating availability
}
\description{
Check if hierarchical C++ implementation is available
}
