% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_observations.R
\name{ChangeObservations}
\alias{ChangeObservations}
\title{Change the observations of fitted Dirichlet Process.}
\usage{
ChangeObservations(dpobj, newData)
}
\arguments{
\item{dpobj}{The Dirichlet process object.}

\item{newData}{New data to be included}
}
\value{
Changed Dirichlet process object
}
\description{
Using a fitted Dirichlet process object include new data. The new data will be assigned to the best fitting cluster for each point.
}
\examples{

y <- rnorm(10)
dp <- DirichletProcessGaussian(y)
dp <- ChangeObservations(dp, rnorm(10))

}
