% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_censored_von_mises.R
\name{circ_in_interval}
\alias{circ_in_interval}
\title{Check if a value is in a circular interval.}
\usage{
circ_in_interval(x, lb, ub)
}
\arguments{
\item{x}{Value to check from 0 to 2pi.}

\item{lb}{Lower bound from 0 to 2pi.}

\item{ub}{Upper bound from 0 to 2pi.}
}
\value{
Logical; whether x lies between lb and ub on the cirle.
}
\description{
Check if a value is in a circular interval.
}
\examples{
circ_in_interval(3, 2, 5)
circ_in_interval(3, 6, 4)
circ_in_interval(3, 6, 2)

}
