% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_component_update.R
\name{ClusterComponentUpdate}
\alias{ClusterComponentUpdate}
\alias{ClusterComponentUpdate.conjugate}
\alias{ClusterComponentUpdate.nonconjugate}
\alias{ClusterComponentUpdate.hierarchical}
\title{Update the component of the Dirichlet process}
\usage{
ClusterComponentUpdate(dpObj)

\method{ClusterComponentUpdate}{conjugate}(dpObj)

\method{ClusterComponentUpdate}{nonconjugate}(dpObj)

\method{ClusterComponentUpdate}{hierarchical}(dpObj)
}
\arguments{
\item{dpObj}{Dirichlet Process object}
}
\value{
Dirichlet process object with update components.
}
\description{
Update the cluster assignment for each data point.
}
\examples{
dp <- DirichletProcessGaussian(rnorm(10))
dp <- ClusterComponentUpdate(dp)

}
