% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_label_change.R
\name{ClusterLabelChange}
\alias{ClusterLabelChange}
\title{Change cluster labels in a Dirichlet Process object}
\usage{
ClusterLabelChange(dpObj, i, newLabel, currentLabel, aux = 0)
}
\arguments{
\item{dpObj}{Dirichlet process object}

\item{i}{Index of the data point to reassign}

\item{newLabel}{New cluster label for the data point}

\item{currentLabel}{Current cluster label of the data point}

\item{aux}{Auxiliary parameters for non-conjugate case}
}
\value{
Updated Dirichlet process object
}
\description{
Internal function to handle cluster label changes, including creation of new clusters
and removal of empty clusters.
}
