% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testHelpers.R
\name{compare_r_cpp}
\alias{compare_r_cpp}
\title{Compare R and C++ function outputs}
\usage{
compare_r_cpp(r_func, cpp_func, ..., tolerance = 1e-10)
}
\arguments{
\item{r_func}{R function to test}

\item{cpp_func}{C++ function to test}

\item{...}{Arguments to pass to both functions}

\item{tolerance}{Numeric tolerance for differences}
}
\value{
Logical indicating whether outputs match
}
\description{
Compare R and C++ function outputs
}
\keyword{internal}
