% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{conjugate_cluster_component_update_cpp}
\alias{conjugate_cluster_component_update_cpp}
\title{Update cluster components (C++ conjugate)}
\usage{
conjugate_cluster_component_update_cpp(dpObj)
}
\arguments{
\item{dpObj}{A list representing the Dirichlet Process object.}
}
\value{
A list with updated cluster assignments and parameters.
}
\description{
C++ implementation of the cluster component update for conjugate models.
}
