% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{conjugate_exponential_cluster_parameter_update_cpp}
\alias{conjugate_exponential_cluster_parameter_update_cpp}
\title{Update cluster parameters (C++ conjugate exponential)}
\usage{
conjugate_exponential_cluster_parameter_update_cpp(dpObj)
}
\arguments{
\item{dpObj}{A list representing the Dirichlet Process object.}
}
\value{
A list containing the updated cluster parameters.
}
\description{
C++ implementation of the cluster parameter update for conjugate models.
}
