% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{conjugate_exponential_update_alpha_cpp}
\alias{conjugate_exponential_update_alpha_cpp}
\title{Update alpha for conjugate exponential DP (C++)}
\usage{
conjugate_exponential_update_alpha_cpp(dpObj)
}
\arguments{
\item{dpObj}{A list representing the Dirichlet Process object.}
}
\value{
Updated alpha value.
}
\description{
C++ implementation of the concentration parameter update.
}
