% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{conjugate_mvnormal_cluster_component_update_cpp}
\alias{conjugate_mvnormal_cluster_component_update_cpp}
\title{Conjugate MVNormal Cluster Component Update (C++)}
\usage{
conjugate_mvnormal_cluster_component_update_cpp(dpObj)
}
\arguments{
\item{dpObj}{Dirichlet process object as list}
}
\value{
Updated Dirichlet process object
}
\description{
Update cluster components for conjugate multivariate normal Dirichlet process
}
