% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{conjugate_mvnormal_cluster_parameter_update_cpp}
\alias{conjugate_mvnormal_cluster_parameter_update_cpp}
\title{Conjugate MVNormal Cluster Parameter Update (C++)}
\usage{
conjugate_mvnormal_cluster_parameter_update_cpp(dpObj)
}
\arguments{
\item{dpObj}{Dirichlet process object as list}
}
\value{
Updated cluster parameters
}
\description{
Update cluster parameters for conjugate multivariate normal Dirichlet process
}
