% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_beta2_exports.R
\name{cpp_beta2_posterior_draw}
\alias{cpp_beta2_posterior_draw}
\title{C++ Beta2 Posterior Draw}
\usage{
cpp_beta2_posterior_draw(data, gamma_prior, maxT, mh_step_size, n, mh_draws)
}
\description{
C++ Beta2 Posterior Draw
}
\keyword{internal}
