% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_hierarchical_beta_wrappers.R
\name{cpp_hierarchical_beta_wrappers}
\alias{cpp_hierarchical_beta_wrappers}
\alias{Fit.hierarchical.cpp}
\alias{ClusterComponentUpdate.hierarchical.cpp}
\alias{GlobalParameterUpdate.hierarchical.cpp}
\alias{UpdateG0.cpp}
\alias{UpdateGamma.cpp}
\alias{HierarchicalBetaCreate.cpp}
\title{C++ Implementation Wrappers for Hierarchical Beta Distribution}
\usage{
\method{Fit}{hierarchical.cpp}(dpObj, its, updatePrior = FALSE, progressBar = interactive())

\method{ClusterComponentUpdate}{hierarchical.cpp}(dpObj)

\method{GlobalParameterUpdate}{hierarchical.cpp}(dpObj)

UpdateG0.cpp(dpObj)

UpdateGamma.cpp(dpObj)

HierarchicalBetaCreate.cpp(
  n,
  priorParameters,
  hyperPriorParameters,
  alphaPrior,
  maxT,
  gammaPrior,
  mhStepSize,
  num_sticks
)
}
\arguments{
\item{dpObj}{Hierarchical Dirichlet process object}

\item{its}{Number of iterations}

\item{updatePrior}{Whether to update prior parameters}

\item{progressBar}{Whether to show progress bar}
}
\description{
These functions provide access to the C++ implementations of the
hierarchical Beta Dirichlet process algorithms.
}
\keyword{internal}
