% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_hierarchical_mvnormal2_wrappers.R
\name{cpp_hierarchical_mvnormal2_wrappers}
\alias{cpp_hierarchical_mvnormal2_wrappers}
\alias{Fit.hierarchical.mvnormal2.cpp}
\alias{HierarchicalMvnormal2Create.cpp}
\alias{ClusterComponentUpdate.mvnormal2.cpp}
\alias{ClusterParameterUpdate.mvnormal2.cpp}
\alias{PriorDraw.mvnormal2.cpp}
\alias{PosteriorDraw.mvnormal2.cpp}
\alias{Likelihood.mvnormal2.cpp}
\title{C++ Implementation Wrappers for Hierarchical MVNormal2 Distribution}
\usage{
\method{Fit}{hierarchical.mvnormal2.cpp}(dpObj, its, updatePrior = FALSE, progressBar = TRUE)

HierarchicalMvnormal2Create.cpp(
  n,
  priorParameters,
  alphaPrior,
  gammaPrior,
  num_sticks
)

\method{ClusterComponentUpdate}{mvnormal2.cpp}(dpObj)

\method{ClusterParameterUpdate}{mvnormal2.cpp}(dpObj)

\method{PriorDraw}{mvnormal2.cpp}(mdObj, n = 1)

\method{PosteriorDraw}{mvnormal2.cpp}(mdObj, x, n = 1, ...)

\method{Likelihood}{mvnormal2.cpp}(mdObj, x, theta)
}
\arguments{
\item{dpObj}{Hierarchical Dirichlet process object}

\item{its}{Number of iterations}

\item{updatePrior}{Whether to update prior parameters}

\item{progressBar}{Whether to show progress bar}
}
\description{
These functions provide access to the C++ implementations of the
hierarchical MVNormal2 Dirichlet process algorithms.
}
\keyword{internal}
