% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_normal_fixed_variance_exports.R
\name{cpp_normal_fixed_variance_likelihood}
\alias{cpp_normal_fixed_variance_likelihood}
\title{C++ Normal Fixed Variance Likelihood}
\usage{
cpp_normal_fixed_variance_likelihood(x, mu, sigma)
}
\description{
C++ Normal Fixed Variance Likelihood
}
\keyword{internal}
