% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_normal_fixed_variance_exports.R
\name{cpp_normal_fixed_variance_posterior_draw}
\alias{cpp_normal_fixed_variance_posterior_draw}
\title{C++ Normal Fixed Variance Posterior Draw}
\usage{
cpp_normal_fixed_variance_posterior_draw(data, mu0, sigma0, sigma, n)
}
\description{
C++ Normal Fixed Variance Posterior Draw
}
\keyword{internal}
