% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_normal_fixed_variance_exports.R
\name{cpp_normal_fixed_variance_posterior_parameters}
\alias{cpp_normal_fixed_variance_posterior_parameters}
\title{C++ Normal Fixed Variance Posterior Parameters}
\usage{
cpp_normal_fixed_variance_posterior_parameters(data, mu0, sigma0, sigma)
}
\description{
C++ Normal Fixed Variance Posterior Parameters
}
\keyword{internal}
