% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_wrappers.R
\name{cpp_normal_wrappers}
\alias{cpp_normal_wrappers}
\alias{normal_prior_draw_cpp_wrapper}
\alias{normal_posterior_draw_cpp_wrapper}
\alias{ClusterComponentUpdate.conjugate.cpp}
\alias{ClusterParameterUpdate.conjugate.cpp}
\title{C++ Implementation Wrappers for Normal Distribution}
\usage{
normal_prior_draw_cpp_wrapper(priorParams, n = 1)

normal_posterior_draw_cpp_wrapper(priorParams, x, n = 1)

\method{ClusterComponentUpdate}{conjugate.cpp}(dpObj)

\method{ClusterParameterUpdate}{conjugate.cpp}(dpObj)
}
\arguments{
\item{priorParams}{Prior parameters (mu0, kappa0, alpha0, beta0)}

\item{n}{Number of draws}

\item{x}{Data matrix}

\item{dpObj}{Dirichlet process object}
}
\description{
These functions provide access to the C++ implementations of the
core sampling algorithms for the Normal distribution.
}
\keyword{internal}
