% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_mcmc_cpp.R
\docType{class}
\name{CppMCMCRunner-class}
\alias{CppMCMCRunner-class}
\alias{CppMCMCRunner}
\title{Manual C++ MCMC Runner Class}
\description{
Manual C++ MCMC Runner Class
}
\section{Fields}{

\describe{
\item{\code{ptr}}{External pointer to C++ MCMCRunnerManual object}

\item{\code{dp_obj}}{Original Dirichlet process object}

\item{\code{distribution_type}}{Type of mixing distribution}
}}

\section{Methods}{

\describe{
\item{\code{get_results()}}{Get accumulated MCMC results}

\item{\code{get_state()}}{Get current MCMC state}

\item{\code{is_complete()}}{Check if all iterations are complete}

\item{\code{run_manual(
  callback = NULL,
  progress = TRUE,
  custom_assignment_sampler = NULL,
  custom_parameter_sampler = NULL
)}}{Run MCMC with optional callbacks and custom samplers}

\item{\code{step()}}{Perform one complete MCMC iteration}

\item{\code{step_assignments()}}{Update cluster assignments using Algorithm 8}

\item{\code{step_concentration()}}{Update concentration parameter}

\item{\code{step_parameters()}}{Update cluster parameters}
}}

