% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_mcmc_cpp.R
\name{create_cpp_mcmc_runner}
\alias{create_cpp_mcmc_runner}
\title{Create Manual C++ MCMC Runner}
\usage{
create_cpp_mcmc_runner(dp_obj, n_iter = 1000, n_burn = 100, thin = 1)
}
\arguments{
\item{dp_obj}{Dirichlet process object (any distribution)}

\item{n_iter}{Number of iterations}

\item{n_burn}{Burn-in iterations}

\item{thin}{Thinning interval}
}
\description{
Create Manual C++ MCMC Runner
}
