% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_mcmc.R
\name{create_gaussian_params}
\alias{create_gaussian_params}
\title{Create Gaussian parameters for debugging}
\usage{
create_gaussian_params()
}
\description{
Create Gaussian parameters for debugging
}
\keyword{internal}
