% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_mcmc.R
\name{debug_mcmc_cpp}
\alias{debug_mcmc_cpp}
\title{Debug MCMC C++ implementation}
\usage{
debug_mcmc_cpp(data, n_iter = 10, verbose = TRUE)
}
\arguments{
\item{data}{Input data for MCMC debugging}

\item{n_iter}{Number of MCMC iterations (default: 10)}

\item{verbose}{Whether to output verbose debugging information (default: TRUE)}
}
\description{
Debug MCMC C++ implementation
}
