% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_mcmc.R
\name{diagnose_clustering}
\alias{diagnose_clustering}
\title{Debug MCMC clustering behavior}
\usage{
diagnose_clustering(data, n_iter = 100, alpha = 1)
}
\arguments{
\item{data}{Input data for clustering diagnosis}

\item{n_iter}{Number of MCMC iterations (default: 100)}

\item{alpha}{Concentration parameter (default: 1.0)}
}
\description{
Debug MCMC clustering behavior
}
