% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_process_gaussian_fixed_variance.R
\name{DirichletProcessGaussianFixedVariance}
\alias{DirichletProcessGaussianFixedVariance}
\title{Create a Dirichlet Mixture of the Gaussian Distribution with fixed variance.}
\usage{
DirichletProcessGaussianFixedVariance(
  y,
  sigma,
  g0Priors = c(0, 1),
  alphaPriors = c(2, 4),
  cpp = FALSE
)
}
\arguments{
\item{y}{Data.}

\item{sigma}{The fixed variance}

\item{g0Priors}{Base Distribution Priors.}

\item{alphaPriors}{Prior parameter distributions for the alpha concentration parameter.}

\item{cpp}{Logical. Use C++ implementation if TRUE, R implementation if FALSE. Default is FALSE.}
}
\value{
Dirichlet process object
}
\description{
Create a Dirichlet Mixture of the Gaussian Distribution with fixed variance.
}
