% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_censored_von_mises.R
\name{DirichletProcessICVonMises}
\alias{DirichletProcessICVonMises}
\title{Create an interval censored Dirichlet Mixture of Von Mises distributions}
\usage{
DirichletProcessICVonMises(y, g0Priors = c(0, 1, 1), alphaPriors = c(2,
  4), muMargMethod = "marginal", n_samp = 3)
}
\arguments{
\item{y}{Data,}

\item{g0Priors}{Base Distribution Priors \eqn{\gamma = (\mu_0, R_0 , n_0)}}

\item{alphaPriors}{Alpha prior parameters. See \code{\link{UpdateAlpha}}.}

\item{muMargMethod}{Method for marginalization of prior mean. See
\code{\link{vonMisesMixtureCreate}}.}

\item{n_samp}{Number of Gibbs samples before we assume draws from posterior
are i.i.d. See \code{\link{vonMisesMixtureCreate}}.}
}
\value{
Dirichlet process object
}
\description{
This is the constructor function to produce a \code{dirichletprocess} object
with a Von Mises mixture kernel for (possibly) interval-censored data with
unknown mean and unknown concentration kappa. The interval censored data is
sampled from the interval according to the current distribution of each
observation dictated by the DPM model. After sampling, the likelihood is simply
conjugate to the base measure and thus the posterior.
}
\details{
The base measure is \eqn{G_0 (\mu, \kappa \mid \gamma) = I_0(R \kappa)^{- n_0}
\exp(R_0 \kappa \cos(\mu - \mu_0))}.
}
