% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive.R
\name{DrawFromDataDistribution}
\alias{DrawFromDataDistribution}
\alias{DrawFromDataDistribution.normal}
\title{Draw data according to the distribution of a dirichletprocess}
\usage{
DrawFromDataDistribution(dpobj, ...)

\method{DrawFromDataDistribution}{normal}(dpobj, params, n = 1, ...)
}
\arguments{
\item{dpobj}{A dirichletprocess object.}

\item{...}{further arguments.}

\item{params}{A parameter vector.}

\item{n}{Number of required samples}
}
\value{
n sampled values.
}
\description{
This function is simply an S3 wrapper around the \code{r____()} function of
each mixing distribution. This construction means that
\code{DrawFromDataDistribution.____} can be easily added for each mixing
distribution, so that we can use \code{PosteriorPredictiveDraw}, for example.
}
\examples{
dp <- Fit(DirichletProcessGaussian(c(rnorm(100), rnorm(100, 3))), 10)
hist(DrawFromDataDistribution(dp, c(1, 5), 1000))

}
