% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_cluster_remove.R
\name{DuplicateClusterRemove}
\alias{DuplicateClusterRemove}
\title{Remove Duplicate Clusters}
\usage{
DuplicateClusterRemove(dpobj)
}
\arguments{
\item{dpobj}{A Dirichlet process object}
}
\value{
Dirichlet process object with duplicate clusters removed
}
\description{
Removes duplicate clusters from a Dirichlet process object.
}
