% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{enable_cpp_hierarchical_samplers}
\alias{enable_cpp_hierarchical_samplers}
\title{Enable C++ implementations for hierarchical models}
\usage{
enable_cpp_hierarchical_samplers(enable = TRUE)
}
\arguments{
\item{enable}{Logical indicating whether to enable hierarchical C++ samplers}
}
\description{
Enable C++ implementations for hierarchical models
}
