% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_wrappers.R
\name{enable_cpp_normal_samplers}
\alias{enable_cpp_normal_samplers}
\title{Enable C++ implementations for Normal samplers}
\usage{
enable_cpp_normal_samplers(use_cpp = TRUE)
}
\arguments{
\item{use_cpp}{Logical indicating whether to use C++ implementations}
}
\description{
This function enables the use of C++ implementations for the Normal
distribution sampling algorithms when available.
}
