% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exponential_likelihood_cpp}
\alias{exponential_likelihood_cpp}
\title{Calculate Exponential likelihood (C++)}
\usage{
exponential_likelihood_cpp(x, lambda)
}
\arguments{
\item{x}{A numeric vector of data points.}

\item{lambda}{The rate parameter.}
}
\value{
A numeric vector of likelihood values.
}
\description{
C++ implementation for calculating exponential likelihood.
}
