% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exponential_posterior_draw_cpp}
\alias{exponential_posterior_draw_cpp}
\title{Draw from an Exponential distribution posterior (C++)}
\usage{
exponential_posterior_draw_cpp(priorParams, x, n = 1L)
}
\arguments{
\item{priorParams}{A numeric vector of prior parameters.}

\item{x}{A numeric matrix of data points.}

\item{n}{The number of samples to draw.}
}
\value{
A list containing the sampled rate parameters (lambda).
}
\description{
C++ implementation for drawing from the posterior distribution of an
  Exponential/Gamma model.
}
