% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exponential_posterior_parameters_cpp}
\alias{exponential_posterior_parameters_cpp}
\title{Calculate Exponential posterior parameters (C++)}
\usage{
exponential_posterior_parameters_cpp(priorParams, x)
}
\arguments{
\item{priorParams}{A numeric vector of prior parameters.}

\item{x}{A numeric matrix of data.}
}
\value{
A list with alpha and beta posterior parameters.
}
\description{
C++ implementation for calculating posterior parameters for an
  Exponential/Gamma model.
}
