% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exponential_predictive_cpp}
\alias{exponential_predictive_cpp}
\title{Calculate Exponential predictive distribution (C++)}
\usage{
exponential_predictive_cpp(priorParams, x)
}
\arguments{
\item{priorParams}{A numeric vector of prior parameters.}

\item{x}{A numeric vector of data.}
}
\value{
A numeric vector of predictive probabilities.
}
\description{
C++ implementation for calculating the predictive distribution.
}
