% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exponential_prior_draw_cpp}
\alias{exponential_prior_draw_cpp}
\title{Draw from an Exponential distribution prior (C++)}
\usage{
exponential_prior_draw_cpp(priorParams, n = 1L)
}
\arguments{
\item{priorParams}{A numeric vector of prior parameters (alpha0, beta0).}

\item{n}{The number of samples to draw.}
}
\value{
A list containing the sampled rate parameters (lambda).
}
\description{
C++ implementation for drawing from the prior distribution of an
  Exponential/Gamma model.
}
