% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_gamma.R
\name{ExponentialMixtureCreate}
\alias{ExponentialMixtureCreate}
\title{Create a Exponential mixing distribution}
\usage{
ExponentialMixtureCreate(priorParameters = c(0.01, 0.01))
}
\arguments{
\item{priorParameters}{Prior parameters for the base measure.}
}
\value{
Mixing distribution object
}
\description{
See \code{\link{DirichletProcessExponential}} for details on the base measure.
}
