% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnormal_normal_wishart.R
\name{extractCovarianceParams}
\alias{extractCovarianceParams}
\title{Extract covariance parameters from matrix}
\usage{
extractCovarianceParams(sigma, covModel)
}
\description{
Extract covariance parameters from matrix
}
\keyword{internal}
