% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_mvnormal_cpp.R
\name{Fit.hdp_mvnormal}
\alias{Fit.hdp_mvnormal}
\title{Fit method for Hierarchical MVNormal DP}
\usage{
\method{Fit}{hdp_mvnormal}(hdp_obj, iterations = 1000, ...)
}
\arguments{
\item{hdp_obj}{Hierarchical DP object}

\item{iterations}{Number of MCMC iterations}

\item{...}{Additional MCMC parameters}
}
\value{
Updated HDP object with samples
}
\description{
Fit method for Hierarchical MVNormal DP
}
