% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hmm.R
\name{fit_hmm}
\alias{fit_hmm}
\title{Fit Hidden Markov Model}
\usage{
fit_hmm(dpObj, its, progressBar = F)
}
\arguments{
\item{dpObj}{Dirichlet Process object}

\item{its}{Number of iterations}

\item{progressBar}{Display progress bar}
}
\value{
Fitted Dirichlet Process object
}
\description{
Internal function for fitting Hidden Markov Dirichlet Process models.
}
