% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{Fit_mvnormal_cpp}
\alias{Fit_mvnormal_cpp}
\title{Specialized Fit function for MVNormal with C++}
\usage{
Fit_mvnormal_cpp(dpObj, its, updatePrior, progressBar)
}
\description{
Specialized Fit function for MVNormal with C++
}
\keyword{internal}
