% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_fixed_variance.R
\name{GaussianFixedVarianceMixtureCreate}
\alias{GaussianFixedVarianceMixtureCreate}
\title{Create a Gaussian Mixing Distribution with fixed variance.}
\usage{
GaussianFixedVarianceMixtureCreate(priorParameters = c(0, 1), sigma)
}
\arguments{
\item{priorParameters}{The prior parameters for the base measure.}

\item{sigma}{The fixed variance of the model.}
}
\value{
A mixing distribution object.
}
\description{
Create a Gaussian Mixing Distribution with fixed variance.
}
