% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diagnostics.R
\name{generate_recommendations}
\alias{generate_recommendations}
\title{Generate MCMC Recommendations}
\usage{
generate_recommendations(gamma_stats, convergence, cluster_stats)
}
\arguments{
\item{gamma_stats}{Gamma chain statistics}

\item{convergence}{Convergence diagnostics}

\item{cluster_stats}{Cluster statistics}
}
\value{
Character vector of recommendations
}
\description{
Generate MCMC Recommendations
}
