% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{get_cpp_status}
\alias{get_cpp_status}
\title{Get C++ implementation status}
\usage{
get_cpp_status()
}
\value{
List showing which C++ implementations are available
}
\description{
Get C++ implementation status
}
