% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_implementation.R
\name{get_implementation}
\alias{get_implementation}
\title{Get Appropriate Implementation}
\usage{
get_implementation(func_name)
}
\arguments{
\item{func_name}{Name of the function to get implementation for}
}
\value{
Function implementation (R or C++ based on setting)
}
\description{
Get Appropriate Implementation
}
\keyword{internal}
