% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnormal_normal_wishart.R
\name{getNumCovParams}
\alias{getNumCovParams}
\title{Get number of covariance parameters for a model}
\usage{
getNumCovParams(d, covModel)
}
\description{
Get number of covariance parameters for a model
}
\keyword{internal}
