% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_parameter_update.R
\name{GlobalParameterUpdate}
\alias{GlobalParameterUpdate}
\title{Update the parameters of the hierarchical Dirichlet process object.}
\usage{
GlobalParameterUpdate(dpobjlist)
}
\arguments{
\item{dpobjlist}{List of Dirichlet Process objects.}
}
\description{
Update the parameters of the hierarchical Dirichlet process object.
}
