% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{has_cpp_implementation}
\alias{has_cpp_implementation}
\title{Check if C++ Implementation is Available}
\usage{
has_cpp_implementation(func_name)
}
\arguments{
\item{func_name}{Name of the function to check}
}
\value{
Logical indicating if the C++ implementation is available
}
\description{
Check if C++ Implementation is Available
}
\keyword{internal}
