% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hierarchical_beta_cluster_component_update_cpp}
\alias{hierarchical_beta_cluster_component_update_cpp}
\title{Update cluster components for Hierarchical Beta DP (C++)}
\usage{
hierarchical_beta_cluster_component_update_cpp(dpList)
}
\arguments{
\item{dpList}{An R list representing the hierarchical DP object.}
}
\value{
Updated hierarchical DP object.
}
\description{
C++ implementation of cluster component update for hierarchical Beta DP.
}
