% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hierarchical_beta_mixing_create_cpp}
\alias{hierarchical_beta_mixing_create_cpp}
\title{Create Hierarchical Beta mixing distributions (C++)}
\usage{
hierarchical_beta_mixing_create_cpp(
  n,
  priorParameters,
  hyperPriorParameters,
  alphaPrior,
  maxT,
  gammaPrior,
  mhStepSize,
  num_sticks
)
}
\arguments{
\item{n}{Number of datasets.}

\item{priorParameters}{Prior parameters for the Beta distribution.}

\item{hyperPriorParameters}{Hyper prior parameters.}

\item{alphaPrior}{Alpha prior parameters.}

\item{maxT}{Maximum value for Beta distribution.}

\item{gammaPrior}{Gamma prior parameters.}

\item{mhStepSize}{Metropolis-Hastings step size.}

\item{num_sticks}{Number of stick breaking values.}
}
\value{
List of mixing distributions.
}
\description{
C++ implementation for creating hierarchical Beta mixing distributions.
}
