% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hierarchical_beta_update_g0_cpp}
\alias{hierarchical_beta_update_g0_cpp}
\title{Update G0 for Hierarchical Beta DP (C++)}
\usage{
hierarchical_beta_update_g0_cpp(dpList)
}
\arguments{
\item{dpList}{An R list representing the hierarchical DP object.}
}
\value{
Updated hierarchical DP object.
}
\description{
C++ implementation of G0 update for hierarchical Beta DP.
}
