% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hierarchical_mvnormal2_fit_cpp}
\alias{hierarchical_mvnormal2_fit_cpp}
\title{Fit Hierarchical MVNormal2 DP (C++)}
\usage{
hierarchical_mvnormal2_fit_cpp(
  dpList,
  iterations,
  updatePrior = FALSE,
  progressBar = TRUE
)
}
\arguments{
\item{dpList}{An R list representing the hierarchical DP object.}

\item{iterations}{Number of iterations.}

\item{updatePrior}{Whether to update prior parameters.}

\item{progressBar}{Whether to show progress bar.}
}
\value{
Updated hierarchical DP object.
}
\description{
C++ implementation for fitting a Hierarchical MVNormal2 DP.
}
