% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hierarchical_mvnormal2_mixing_create_cpp}
\alias{hierarchical_mvnormal2_mixing_create_cpp}
\title{Create Hierarchical MVNormal2 mixing distributions (C++)}
\usage{
hierarchical_mvnormal2_mixing_create_cpp(
  n,
  priorParameters,
  alphaPrior,
  gammaPrior,
  num_sticks
)
}
\arguments{
\item{n}{Number of datasets.}

\item{priorParameters}{Prior parameters for the MVNormal2 distribution.}

\item{alphaPrior}{Alpha prior parameters.}

\item{gammaPrior}{Gamma prior parameters.}

\item{num_sticks}{Number of stick breaking values.}
}
\value{
List of mixing distributions.
}
\description{
C++ implementation for creating hierarchical MVNormal2 mixing distributions.
}
