% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hierarchical_mvnormal_create_mixing}
\alias{hierarchical_mvnormal_create_mixing}
\title{Create Hierarchical MVNormal mixing distributions (C++)}
\usage{
hierarchical_mvnormal_create_mixing(
  n_groups,
  prior_params,
  alpha_prior,
  gamma_prior,
  n_sticks
)
}
\arguments{
\item{n_groups}{Number of groups}

\item{prior_params}{Prior parameters for base distribution}

\item{alpha_prior}{Prior for local concentration parameters}

\item{gamma_prior}{Prior for global concentration parameter}

\item{n_sticks}{Number of stick-breaking components}
}
\value{
List representing the mixing distribution
}
\description{
Initialize hierarchical MVNormal mixing structure
}
