% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hierarchical_mvnormal_fit_cpp}
\alias{hierarchical_mvnormal_fit_cpp}
\title{Fit Hierarchical MVNormal DP (C++)}
\usage{
hierarchical_mvnormal_fit_cpp(
  dp_list,
  iterations,
  update_prior = TRUE,
  progress_bar = TRUE
)
}
\arguments{
\item{dp_list}{List of DP objects for each group}

\item{iterations}{Number of MCMC iterations}

\item{update_prior}{Whether to update hyperparameters}

\item{progress_bar}{Whether to show progress}
}
\value{
Updated hierarchical DP object
}
\description{
Complete fitting routine for hierarchical MVNormal DP
}
