% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hierarchical_mvnormal_posterior_sample}
\alias{hierarchical_mvnormal_posterior_sample}
\title{Sample from hierarchical MVNormal posterior (C++)}
\usage{
hierarchical_mvnormal_posterior_sample(hdp_state, n_samples, group_index)
}
\arguments{
\item{hdp_state}{Current state of the hierarchical DP}

\item{n_samples}{Number of samples to draw}

\item{group_index}{Which group to sample for (0-indexed)}
}
\value{
Matrix of samples
}
\description{
Draw samples from the posterior predictive distribution
}
