% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hierarchical_mvnormal_update_clusters}
\alias{hierarchical_mvnormal_update_clusters}
\title{Update cluster assignments for Hierarchical MVNormal (C++)}
\usage{
hierarchical_mvnormal_update_clusters(dp_obj, global_params)
}
\arguments{
\item{dp_obj}{Dirichlet process object for a single group}

\item{global_params}{Current global parameters}
}
\value{
Updated DP object
}
\description{
Update cluster assignments using Algorithm 8 for a single group
}
