% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_beta.R
\name{HierarchicalBetaCreate}
\alias{HierarchicalBetaCreate}
\title{Create a Mixing Object for a hierarchical Beta Dirichlet process object.}
\usage{
HierarchicalBetaCreate(n, priorParameters, hyperPriorParameters, alpha0, maxT,
  gammaParam, mhStepSize, num_sticks)
}
\arguments{
\item{n}{Number of data sets}

\item{priorParameters}{The prior parameters for the top level base distribution.}

\item{hyperPriorParameters}{Hyper prior parameters for the top level base distribution.}

\item{alpha0}{Individual level concentration parameter.}

\item{maxT}{Bounding value of the data.}

\item{gammaParam}{Concentration parameter for the top level.}

\item{mhStepSize}{Metropolis Hastings step size for the posterior drawing.}

\item{num_sticks}{Number of stick breaking values to use.}
}
\value{
A mixing distribution object.
}
\description{
Create a Mixing Object for a hierarchical Beta Dirichlet process object.
}
