% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_mvnormal_cpp.R
\name{HierarchicalDirichletProcessMVNormal}
\alias{HierarchicalDirichletProcessMVNormal}
\title{Create Hierarchical MVNormal Dirichlet Process}
\usage{
HierarchicalDirichletProcessMVNormal(
  data_list,
  prior_params,
  alpha_prior = c(1, 1),
  gamma_prior = c(1, 1),
  n_sticks = 20
)
}
\arguments{
\item{data_list}{List of data matrices}

\item{prior_params}{Prior parameters for MVNormal-Wishart}

\item{alpha_prior}{Prior for local concentration parameters}

\item{gamma_prior}{Prior for global concentration parameter}

\item{n_sticks}{Number of stick-breaking components}
}
\value{
Hierarchical DP object
}
\description{
Create Hierarchical MVNormal Dirichlet Process
}
