% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_mvnormal2.R
\name{HierarchicalMvnormal2Create}
\alias{HierarchicalMvnormal2Create}
\title{Create a Mixing Object for a hierarchical semi-conjugate
Multivariate Normal Dirichlet process object.}
\usage{
HierarchicalMvnormal2Create(
  n,
  priorParameters,
  alphaPrior,
  gammaPrior,
  num_sticks
)
}
\arguments{
\item{n}{Number of data sets}

\item{priorParameters}{The prior parameters for the top level base distribution.}

\item{alphaPrior}{Individual level concentration parameter priors.}

\item{gammaPrior}{Concentration parameter for the top level priors.}

\item{num_sticks}{Number of stick breaking values to use.}
}
\value{
A mixing distribution object.
}
\description{
Create a Mixing Object for a hierarchical semi-conjugate
Multivariate Normal Dirichlet process object.
}
