% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_distribution_likelihood.R
\name{Likelihood}
\alias{Likelihood}
\title{Mixing Distribution Likelihood}
\usage{
Likelihood(mdObj, x, theta)
}
\arguments{
\item{mdObj}{Mixing Distribution}

\item{x}{Data}

\item{theta}{Parameters of distribution}
}
\value{
Likelihood of the data
}
\description{
Evaluate the Likelihood of some data \eqn{x} for some parameter \eqn{\theta}.
}
