% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{likelihood_cpp_wrapper}
\alias{likelihood_cpp_wrapper}
\title{Interface to C++ Likelihood Implementation}
\usage{
likelihood_cpp_wrapper(mdObj, x, theta)
}
\arguments{
\item{mdObj}{Mixing distribution object}

\item{x}{Data vector/matrix}

\item{theta}{List of parameters}
}
\value{
Vector of likelihood values
}
\description{
This function provides a bridge between the R S3 methods and the C++ implementation
of the likelihood functions.
}
\keyword{internal}
