% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{LikelihoodDP}
\alias{LikelihoodDP}
\title{The likelihood of the Dirichlet process object}
\usage{
LikelihoodDP(dpobj)
}
\arguments{
\item{dpobj}{The dirichletprocess object on which to calculate the likelihood.}
}
\description{
Calculate the likelihood of each data point with its parameter.
}
