% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{LikelihoodFunction}
\alias{LikelihoodFunction}
\title{The Likelihood of a Dirichlet process object.}
\usage{
LikelihoodFunction(dpobj)
}
\arguments{
\item{dpobj}{Dirichlet process object.}
}
\value{
A function f(x) that represents the Likelihood of the dpobj.
}
\description{
Collecting the fitted cluster parameters and number of datapoints associated with each parameter a Likelihood can be calculated.
Each cluster is weighted by the number of datapoints assigned.
}
\examples{
y <- rnorm(10)
dp <- DirichletProcessGaussian(y)
dp <- Fit(dp, 5)
f <- LikelihoodFunction(dp)
plot(f(-2:2))

}
