% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{markov_dp_create_cpp}
\alias{markov_dp_create_cpp}
\title{Create a Markov DP from R object (C++)}
\usage{
markov_dp_create_cpp(dpObj)
}
\arguments{
\item{dpObj}{An R list representing the Markov DP object.}
}
\value{
An updated list with C++ object reference.
}
\description{
C++ implementation for creating a Markov DP from an R object.
}
