% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{markov_dp_fit_cpp}
\alias{markov_dp_fit_cpp}
\title{Fit Markov DP (C++)}
\usage{
markov_dp_fit_cpp(dpObj, iterations, updatePrior = FALSE, progressBar = TRUE)
}
\arguments{
\item{dpObj}{An R list representing the Markov DP object.}

\item{iterations}{Number of iterations.}

\item{updatePrior}{Whether to update prior parameters.}

\item{progressBar}{Whether to show progress bar.}
}
\value{
Updated Markov DP object.
}
\description{
C++ implementation for fitting a Markov DP (HMM).
}
