% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{markov_dp_param_update_cpp}
\alias{markov_dp_param_update_cpp}
\title{Update parameters for Markov DP (C++)}
\usage{
markov_dp_param_update_cpp(dpObj)
}
\arguments{
\item{dpObj}{An R list representing the Markov DP object.}
}
\value{
Updated Markov DP object.
}
\description{
C++ implementation of parameter update for Markov DP.
}
