% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{markov_dp_update_states_cpp}
\alias{markov_dp_update_states_cpp}
\title{Update states for Markov DP (C++)}
\usage{
markov_dp_update_states_cpp(dpObj)
}
\arguments{
\item{dpObj}{An R list representing the Markov DP object.}
}
\value{
Updated Markov DP object.
}
\description{
C++ implementation of state update for Markov DP.
}
