% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis_hastings.R
\name{MetropolisHastings}
\alias{MetropolisHastings}
\title{Metropolis-Hastings MCMC Sampler}
\usage{
MetropolisHastings(mixingDistribution, x, start_pos, no_draws = 100)
}
\arguments{
\item{mixingDistribution}{A mixing distribution object}

\item{x}{Data for which to sample parameters}

\item{start_pos}{Starting position for the MCMC chain - a list of parameter arrays}

\item{no_draws}{Number of MCMC draws to perform (default: 100)}
}
\value{
A list containing:
\itemize{
  \item parameter_samples: List of parameter sample arrays
  \item accept_ratio: Acceptance ratio of the MCMC chain
}
}
\description{
Performs Metropolis-Hastings sampling for non-conjugate Dirichlet process mixtures.
This function is used internally for parameter updates in non-conjugate models
where analytical posterior updates are not available.
}
\details{
This function implements the Metropolis-Hastings algorithm for sampling
from posterior distributions in non-conjugate Dirichlet process mixtures.
Different mixing distributions may have specialized implementations.
}
\references{
Metropolis, N., et al. (1953). Equation of state calculations by fast computing machines.
            Journal of Chemical Physics, 21(6), 1087-1092.
}
