% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_distribution.R
\name{MixingDistribution}
\alias{MixingDistribution}
\title{Create a mixing distribution object}
\usage{
MixingDistribution(
  distribution,
  priorParameters,
  conjugate,
  mhStepSize = NULL,
  hyperPriorParameters = NULL
)
}
\arguments{
\item{distribution}{The name of the distribution mixture}

\item{priorParameters}{The prior parameters}

\item{conjugate}{Whether the prior is conjugate to the Likelihood.}

\item{mhStepSize}{The scale of the proposal parameter for the Metropolis Hastings algorithm. Not needed for conjugate mixtures.}

\item{hyperPriorParameters}{Vector of hyperPriorParameters for the distribution.}
}
\description{
The constructor function for a mixing distribution object.
Use this function to prepare an object for use with the appropriate distribution functions.
}
