% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvnormal2_likelihood_cpp}
\alias{mvnormal2_likelihood_cpp}
\title{Calculate MVNormal2 likelihood (C++)}
\usage{
mvnormal2_likelihood_cpp(x, theta)
}
\arguments{
\item{x}{A numeric vector of a single data point.}

\item{theta}{A list containing mu and sig parameters.}
}
\value{
A numeric vector of likelihood values.
}
\description{
C++ implementation for calculating multivariate normal likelihood.
}
