% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvnormal2_prior_draw_cpp}
\alias{mvnormal2_prior_draw_cpp}
\title{Draw from a Multivariate Normal semi-conjugate prior (C++)}
\usage{
mvnormal2_prior_draw_cpp(priorParams, n = 1L)
}
\arguments{
\item{priorParams}{A list containing prior parameters (mu0, sigma0, phi0, nu0).}

\item{n}{The number of samples to draw.}
}
\value{
A list containing the sampled parameters (mu and sig).
}
\description{
C++ implementation for drawing from the prior distribution of a
  Multivariate Normal semi-conjugate model.
}
