% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnormal_semi_conjugate.R
\name{Mvnormal2Create}
\alias{Mvnormal2Create}
\title{Create a multivariate normal mixing distribution with semi conjugate prior}
\usage{
Mvnormal2Create(priorParameters)
}
\arguments{
\item{priorParameters}{The prior parameters for the Multivariate Normal.}
}
\description{
Create a multivariate normal mixing distribution with semi conjugate prior
}
