% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvnormal_likelihood_cpp}
\alias{mvnormal_likelihood_cpp}
\title{Calculate MVNormal likelihood (C++)}
\usage{
mvnormal_likelihood_cpp(x, mu, sigma)
}
\arguments{
\item{x}{A numeric matrix of data points.}

\item{mu}{Mean vector.}

\item{sigma}{Covariance matrix.}
}
\value{
A numeric vector of likelihood values.
}
\description{
C++ implementation for calculating multivariate normal likelihood.
}
