% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnormal_normal_wishart.R
\name{mvnormal_likelihood_wrapper_cpp}
\alias{mvnormal_likelihood_wrapper_cpp}
\title{C++ wrapper for likelihood calculation}
\usage{
mvnormal_likelihood_wrapper_cpp(x, theta, priorParams)
}
\description{
C++ wrapper for likelihood calculation
}
\keyword{internal}
