% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvnormal_posterior_draw_cpp}
\alias{mvnormal_posterior_draw_cpp}
\title{Draw from a Multivariate Normal-Wishart posterior (C++)}
\usage{
mvnormal_posterior_draw_cpp(priorParams, x, n = 1L)
}
\arguments{
\item{priorParams}{A list containing prior parameters.}

\item{x}{A numeric matrix of data points.}

\item{n}{The number of samples to draw.}
}
\value{
A list containing the sampled parameters (mu and sig).
}
\description{
C++ implementation for drawing from the posterior distribution of a
  Multivariate Normal-Wishart model.
}
