% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvnormal_posterior_parameters_cpp}
\alias{mvnormal_posterior_parameters_cpp}
\title{Calculate MVNormal posterior parameters (C++)}
\usage{
mvnormal_posterior_parameters_cpp(priorParams, x)
}
\arguments{
\item{priorParams}{A list containing prior parameters.}

\item{x}{A numeric matrix of data.}
}
\value{
A list of posterior parameters.
}
\description{
C++ implementation for calculating posterior parameters for a
  Multivariate Normal-Wishart model.
}
