% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvnormal_predictive_cpp}
\alias{mvnormal_predictive_cpp}
\title{Calculate MVNormal predictive distribution (C++)}
\usage{
mvnormal_predictive_cpp(priorParams, x)
}
\arguments{
\item{priorParams}{A list containing prior parameters.}

\item{x}{A numeric matrix of data.}
}
\value{
A numeric vector of predictive probabilities.
}
\description{
C++ implementation for calculating the predictive distribution.
}
