% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvnormal_prior_draw_cpp}
\alias{mvnormal_prior_draw_cpp}
\title{Draw from a Multivariate Normal-Wishart prior (C++)}
\usage{
mvnormal_prior_draw_cpp(priorParams, n = 1L)
}
\arguments{
\item{priorParams}{A list containing prior parameters (mu0, kappa0, Lambda, nu).}

\item{n}{The number of samples to draw.}
}
\value{
A list containing the sampled parameters (mu and sig).
}
\description{
C++ implementation for drawing from the prior distribution of a
  Multivariate Normal-Wishart model.
}
