% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nonconjugate_mvnormal2_cluster_component_update_cpp}
\alias{nonconjugate_mvnormal2_cluster_component_update_cpp}
\title{Update cluster components for MVNormal2 (C++ non-conjugate)}
\usage{
nonconjugate_mvnormal2_cluster_component_update_cpp(dpObj)
}
\arguments{
\item{dpObj}{A list representing the Dirichlet Process object.}
}
\value{
A list with updated cluster assignments and parameters.
}
\description{
C++ implementation of the cluster component update for MVNormal2 non-conjugate models.
}
