% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nonconjugate_mvnormal2_cluster_parameter_update_cpp}
\alias{nonconjugate_mvnormal2_cluster_parameter_update_cpp}
\title{Update cluster parameters for MVNormal2 (C++ non-conjugate)}
\usage{
nonconjugate_mvnormal2_cluster_parameter_update_cpp(dpObj)
}
\arguments{
\item{dpObj}{A list representing the Dirichlet Process object.}
}
\value{
A list containing the updated cluster parameters.
}
\description{
C++ implementation of the cluster parameter update for MVNormal2 non-conjugate models.
}
