% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nonconjugate_mvnormal2_update_alpha_cpp}
\alias{nonconjugate_mvnormal2_update_alpha_cpp}
\title{Update alpha for non-conjugate MVNormal2 DP (C++)}
\usage{
nonconjugate_mvnormal2_update_alpha_cpp(dpObj)
}
\arguments{
\item{dpObj}{A list representing the Dirichlet Process object.}
}
\value{
Updated alpha value.
}
\description{
C++ implementation of the concentration parameter update for MVNormal2.
}
