% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normal_posterior_parameters_cpp}
\alias{normal_posterior_parameters_cpp}
\title{Calculate Normal posterior parameters (C++)}
\usage{
normal_posterior_parameters_cpp(priorParams, x)
}
\arguments{
\item{priorParams}{A numeric vector of prior parameters.}

\item{x}{A numeric matrix of data.}
}
\value{
A numeric matrix of posterior parameters.
}
\description{
C++ implementation for calculating posterior parameters for a
  Normal/Inverse-Gamma model.
}
