% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normal_prior_draw_cpp}
\alias{normal_prior_draw_cpp}
\title{Draw from a Normal distribution prior (C++)}
\usage{
normal_prior_draw_cpp(priorParams, n = 1L)
}
\arguments{
\item{priorParams}{A numeric vector of prior parameters.}

\item{n}{The number of samples to draw.}
}
\value{
A list containing the sampled parameters (mu and sigma^2).
}
\description{
C++ implementation for drawing from the prior distribution of a
  Normal/Inverse-Gamma model.
}
