% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_uniform_gamma.R,
%   R/mixing_distribution_penalised_likelihood.R
\name{PenalisedLikelihood.beta}
\alias{PenalisedLikelihood.beta}
\alias{PenalisedLikelihood}
\alias{PenalisedLikelihood.default}
\title{Calculate the parameters that maximise the penalised likelihood.}
\usage{
\method{PenalisedLikelihood}{beta}(mdObj, x)

PenalisedLikelihood(mdObj, x)

\method{PenalisedLikelihood}{default}(mdObj, x)
}
\arguments{
\item{mdObj}{Mixing distribution object}

\item{x}{Data}
}
\description{
Used to find suitable starting parameters for nonconjugate mixtures.
For some mixing distributions this hasn't been implemented yet.
}
