% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_clusters.R,
%   R/zzz_hierarchical_phase5_helpers.R
\name{PosteriorClusters}
\alias{PosteriorClusters}
\alias{PosteriorClusters.hierarchical}
\title{Draw posterior clusters conditional on one fitted state}
\usage{
PosteriorClusters(dpobj, ind)

\method{PosteriorClusters}{hierarchical}(dpobj, ind)
}
\arguments{
\item{dpobj}{Fitted Dirichlet process}

\item{ind}{Stored-iteration index for which the posterior will be drawn.
If omitted, the current fitted state is used.}
}
\value{
A list with the weights and cluster parameters that form a
  conditional posterior draw from the Dirichlet process.
}
\description{
Using the stick-breaking representation, this lower-level helper draws
posterior clusters and weights conditional on the current fitted state, or
on one retained stored iteration when \code{ind} is supplied. See also
\code{\link{PosteriorFunction}} and \code{\link{PosteriorSummary}}.
}
\examples{
y <- rnorm(10)
dp <- DirichletProcessGaussian(y)
dp <- Fit(dp, 5, progressBar = FALSE)
postClusters <- PosteriorClusters(dp)

}
