% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R,
%   R/zzz_hierarchical_phase5_helpers.R
\name{PosteriorFunction}
\alias{PosteriorFunction}
\alias{PosteriorFunction.hierarchical}
\title{Draw a posterior function conditional on one fitted state}
\usage{
PosteriorFunction(dpobj, ind)

\method{PosteriorFunction}{hierarchical}(dpobj, ind)
}
\arguments{
\item{dpobj}{Fitted Dirichlet Process object}

\item{ind}{Stored-iteration index to draw from. If omitted, the current
fitted state is used.}
}
\value{
A posterior function \code{f(x)}.
}
\description{
Lower-level helper for drawing a posterior function conditional on the
current fitted state, or on one retained stored iteration when
\code{ind} is supplied. For retained-sample posterior summaries across
multiple stored iterations, prefer \code{\link{PosteriorSummary}}.
}
\examples{

y <- rnorm(10)
dp <- DirichletProcessGaussian(y)
dp <- Fit(dp, 5, progressBar = FALSE)
postFuncDraw <- PosteriorFunction(dp)
plot(-3:3, postFuncDraw(-3:3))

}
