% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive.R
\name{PosteriorPredictiveDraw}
\alias{PosteriorPredictiveDraw}
\title{Draw from the posterior predictive distribution of a dirichlet process model.}
\usage{
PosteriorPredictiveDraw(dpobj, ...)
}
\arguments{
\item{dpobj}{A dirichletprocess object.}
}
\value{
A numeric vector of samples from the posterior predictive
  distribution.
}
\description{
By drawing from the posterior predictive distribution, we generate possible
future observations, taking into account the uncertainty in the parameters.
Thus, this function is useful for prediction as well as decision making.
}
\examples{
dp <- Fit(DirichletProcessGaussian(c(rnorm(100), rnorm(100, 3))), 10)
PosteriorPredictiveDraw(dp, 10)

}
