% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive.R
\name{PosteriorPredictiveInterval}
\alias{PosteriorPredictiveInterval}
\alias{PosteriorPredictiveInterval.dirichletprocess}
\title{Obtain an estimate of the proportion of future data in a certain interval}
\usage{
PosteriorPredictiveInterval(dpobj, ...)

\method{PosteriorPredictiveInterval}{dirichletprocess}(dpobj, from, to,
  n_reps = 100, ci_size = 0.05, ...)
}
\arguments{
\item{dpobj}{A dirichletprocess object.}

\item{...}{further arguments.}

\item{from}{Start of the interval.}

\item{to}{End of the interval.}

\item{n_reps}{Nnumber of posterior samples to use to estimate the interval.}

\item{ci_size}{Width of the credible interval around the probability
estimate.}
}
\value{
The median probabiliity along with the credible interval.
}
\description{
Using the posterior predictive distribution, we can obtain an estimate of the
proportion of data in a certain interval, as well as the uncertainty around
this estimate across the posterior.
}
\examples{
dp <- Fit(DirichletProcessGaussian(c(rnorm(100), rnorm(100, 3))), 10)
PosteriorPredictiveInterval(dp, 0, 3)

}
