% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_cpp_interface.R
\name{prepare_markov_mixing_params}
\alias{prepare_markov_mixing_params}
\title{Prepare Markov mixing distribution parameters for C++}
\usage{
prepare_markov_mixing_params(md)
}
\arguments{
\item{md}{Mixing distribution object}
}
\value{
List of parameters for C++
}
\description{
Prepare Markov mixing distribution parameters for C++
}
\keyword{internal}
