% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{prepare_mixing_dist_params}
\alias{prepare_mixing_dist_params}
\title{Create mixing distribution parameters for C++}
\usage{
prepare_mixing_dist_params(dp_obj)
}
\arguments{
\item{dp_obj}{Dirichlet process object}
}
\value{
List of parameters formatted for C++
}
\description{
Create mixing distribution parameters for C++
}
\keyword{internal}
