% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diagnostics.R
\name{print.hdp_diagnostics}
\alias{print.hdp_diagnostics}
\title{Print method for HDP diagnostics}
\usage{
\method{print}{hdp_diagnostics}(x, ...)
}
\arguments{
\item{x}{An hdp_diagnostics object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for HDP diagnostics
}
