% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_distribution_prior_density.R
\name{PriorDensity}
\alias{PriorDensity}
\title{Calculate the prior density of a mixing distribution}
\usage{
PriorDensity(mdObj, x)
}
\arguments{
\item{mdObj}{Mixing distribution}

\item{x}{Prior parameters}
}
\description{
Calculate the prior density of a mixing distribution
}
