% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnormal_normal_wishart.R
\name{reconstructCovarianceMatrix}
\alias{reconstructCovarianceMatrix}
\title{Reconstruct covariance matrix from parameters}
\usage{
reconstructCovarianceMatrix(params, d, covModel)
}
\description{
Reconstruct covariance matrix from parameters
}
\keyword{internal}
