% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{register_cpp_implementations}
\alias{register_cpp_implementations}
\title{Register C++ Implementations}
\usage{
register_cpp_implementations()
}
\description{
This function registers the C++ implementations as alternatives to the R implementations.
It should be called during package loading.
}
\keyword{internal}
