% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_mvnormal_cpp.R
\name{run_hierarchical_mvnormal_mcmc_cpp}
\alias{run_hierarchical_mvnormal_mcmc_cpp}
\title{Run Hierarchical MVNormal MCMC using C++}
\usage{
run_hierarchical_mvnormal_mcmc_cpp(data_list, hdp_params, mcmc_params)
}
\arguments{
\item{data_list}{List of data matrices}

\item{hdp_params}{List of HDP parameters}

\item{mcmc_params}{List of MCMC parameters}
}
\value{
List with MCMC results
}
\description{
Run Hierarchical MVNormal MCMC using C++
}
