% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_cpp_interface.R
\name{run_markov_mcmc_cpp}
\alias{run_markov_mcmc_cpp}
\title{Run Markov MCMC using C++ implementation}
\usage{
run_markov_mcmc_cpp(dp_obj, its, update_prior = FALSE, progress_bar = TRUE)
}
\arguments{
\item{dp_obj}{Markov Dirichlet process object}

\item{its}{Number of iterations}

\item{update_prior}{Whether to update hyperparameters}

\item{progress_bar}{Whether to show progress}
}
\value{
Updated DP object
}
\description{
Run Markov MCMC using C++ implementation
}
\keyword{internal}
