% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_censored.R
\name{rvm_ic}
\alias{rvm_ic}
\title{Sample from the von Mises distribution with interval constraints.}
\usage{
rvm_ic(n, mu = 0, kp = 1, lb = 0, ub = 2 * pi,
  method = "adaptive", adaptive_cutoff = 0.1, max_attempts = 1000)
}
\arguments{
\item{n}{Number of required samples.}

\item{mu}{Mean direction of the von Mises distribution.}

\item{kp}{Concentration parameter of the von Mises distribution.}

\item{lb}{Lower bound.}

\item{ub}{Upper bound.}
}
\value{
A numeric vector of sampled values.
}
\description{
Sample from the von Mises distribution with interval constraints.
}
\examples{
x <- rvm_ic(10000, 0, 4, 5, 1)
hist(x, breaks = 100)
}
