% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_cpp_interface.R
\name{update_dp_from_mcmc}
\alias{update_dp_from_mcmc}
\title{Update DP object from MCMC results}
\usage{
update_dp_from_mcmc(dp, mcmc_result)
}
\arguments{
\item{dp}{Original DP object}

\item{mcmc_result}{MCMC results from C++}
}
\value{
Updated DP object
}
\description{
Update DP object from MCMC results
}
\keyword{internal}
