% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_alpha_beta.R
\name{UpdateAlphaBeta}
\alias{UpdateAlphaBeta}
\title{Update the $\alpha$ and $\beta$ parameter of a hidden Markov Dirichlet process model.}
\usage{
UpdateAlphaBeta(dp)
}
\description{
Update the $\alpha$ and $\beta$ parameter of a hidden Markov Dirichlet process model.
}
