% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_states.R
\name{UpdateStates}
\alias{UpdateStates}
\title{Update States for Dirichlet Process}
\usage{
UpdateStates(dp)
}
\arguments{
\item{dp}{A Dirichlet process object}
}
\value{
Updated Dirichlet process object
}
\description{
Updates the states of a Dirichlet process object using either C++ or R implementation.
}
