% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R, R/switch_implementation.R,
%   R/zzz.R
\name{using_cpp}
\alias{using_cpp}
\title{Check if C++ implementations are enabled}
\usage{
using_cpp()

using_cpp()

using_cpp()
}
\value{
Logical indicating if C++ implementations are enabled

Logical indicating if C++ implementations are being used
}
\description{
Check whether R or C++ implementations are being used
}
