% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R, R/cpp_wrappers.R
\name{using_cpp_samplers}
\alias{using_cpp_samplers}
\title{Check if cpp samplers are enabled (general)}
\usage{
using_cpp_samplers()

using_cpp_samplers()
}
\value{
Logical indicating if C++ samplers are enabled
}
\description{
Check if cpp samplers are enabled (general)

Check if C++ samplers are enabled
}
\keyword{internal}
