% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_censored_von_mises.R
\name{vonMisesICMixtureCreate}
\alias{vonMisesICMixtureCreate}
\title{Create a von Mises interval censored mixing distribution}
\usage{
vonMisesICMixtureCreate(priorParameters, muMargSample = "marginal")
}
\arguments{
\item{priorParameters}{Prior parameters for the base measure which are, in
order, (mu_0, R_0, c).}
}
\value{
Mixing distribution object
}
\description{
Create a von Mises interval censored mixing distribution
}
