% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/von_mises_bessel_exponential.R
\name{vonMisesMixtureCreate}
\alias{vonMisesMixtureCreate}
\title{Create a von Mises mixing distribution}
\usage{
vonMisesMixtureCreate(priorParameters)
}
\arguments{
\item{priorParameters}{Prior parameters for the base measure which are, in
order, (mu_0, R_0, c).}
}
\value{
Mixing distribution object
}
\description{
Create a von Mises mixing distribution
}
