% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_uniform_gamma.R
\name{WeibullMixtureCreate}
\alias{WeibullMixtureCreate}
\title{Create a Weibull mixing distribution.}
\usage{
WeibullMixtureCreate(priorParameters, mhStepSize, hyperPriorParameters = c(6,
  2, 1, 0.5))
}
\arguments{
\item{priorParameters}{Prior parameters for the Weibull parameters}

\item{mhStepSize}{Metropolis Hastings Step Size}

\item{hyperPriorParameters}{Parameters for the hyper-priors}
}
\value{
A mixing distribution object.
}
\description{
See \code{\link{DirichletProcessWeibull}} for the default prior and hyper prior distributions.
}
