% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{weighted_function_generator}
\alias{weighted_function_generator}
\title{Generate a weighted function.}
\usage{
weighted_function_generator(func, weights, params)
}
\arguments{
\item{func}{Function that is used of the form func(x, params).}

\item{weights}{Weighting of each cluster.}

\item{params}{Cluster parameter list}
}
\value{
weighted function
}
\description{
Generate a weighted function.
}
