% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_uniform_gamma.R
\name{BetaMixtureCreate}
\alias{BetaMixtureCreate}
\title{Create a Beta mixing distribution.}
\usage{
BetaMixtureCreate(priorParameters = c(2, 8), mhStepSize = c(1, 1),
  maxT = 1, hyperPriorParameters = c(1, 0.125))
}
\arguments{
\item{priorParameters}{The prior parameters for the base measure.}

\item{mhStepSize}{The Metropolis Hastings step size. A numeric vector of length 2.}

\item{maxT}{The upper bound of the Beta distribution. Defaults to 1 for the standard Beta distribution.}

\item{hyperPriorParameters}{The parameters for the hyper prior.}
}
\value{
A mixing distribution object.
}
\description{
See \code{\link{DirichletProcessBeta}} for the default prior and hyper prior distributions.
}
