% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{can_use_cpp}
\alias{can_use_cpp}
\title{Check if C++ can be used for a given DP object}
\usage{
can_use_cpp(dp_obj = NULL)
}
\arguments{
\item{dp_obj}{Dirichlet process object}
}
\value{
Logical indicating whether C++ implementation is available
}
\description{
Check if C++ can be used for a given DP object
}
\keyword{internal}
