% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_markov_wrappers.R
\name{cpp_markov_wrappers}
\alias{cpp_markov_wrappers}
\alias{Fit.markov.cpp}
\alias{UpdateStates.cpp}
\alias{UpdateAlphaBeta.cpp}
\alias{param_update.cpp}
\title{C++ Implementation Wrappers for Markov DP (HMM)}
\usage{
\method{Fit}{markov.cpp}(dpObj, its, updatePrior = FALSE, progressBar = TRUE, ...)

UpdateStates.cpp(dpObj)

UpdateAlphaBeta.cpp(dpObj)

param_update.cpp(dpObj)
}
\arguments{
\item{dpObj}{Markov Dirichlet process object}

\item{its}{Number of iterations}

\item{progressBar}{Whether to show progress bar}
}
\description{
These functions provide access to the C++ implementations of the
Markov Dirichlet process (Hidden Markov Model) algorithms.
}
\keyword{internal}
