% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_mvnormal_wrappers.R
\name{cpp_mvnormal_wrappers}
\alias{cpp_mvnormal_wrappers}
\alias{ClusterComponentUpdate.mvnormal.cpp}
\alias{ClusterParameterUpdate.mvnormal.cpp}
\alias{PriorDraw.mvnormal.cpp}
\alias{PosteriorDraw.mvnormal.cpp}
\alias{Likelihood.mvnormal.cpp}
\title{C++ Implementation Wrappers for MVNormal Distribution}
\usage{
\method{ClusterComponentUpdate}{mvnormal.cpp}(dpObj)

\method{ClusterParameterUpdate}{mvnormal.cpp}(dpObj)

\method{PriorDraw}{mvnormal.cpp}(mdObj, n = 1, ...)

\method{PosteriorDraw}{mvnormal.cpp}(mdObj, x, n = 1, ...)

\method{Likelihood}{mvnormal.cpp}(mdObj, x, theta)
}
\arguments{
\item{dpObj}{Dirichlet process object}

\item{n}{Number of draws}

\item{x}{Data matrix}
}
\description{
These functions provide access to the C++ implementations of the
core sampling algorithms for the MVNormal distribution.
}
\keyword{internal}
