% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_mcmc_cpp.R
\docType{class}
\name{CppMCMCRunner-class}
\alias{CppMCMCRunner-class}
\alias{CppMCMCRunner}
\title{Manual C++ MCMC Runner Class}
\description{
Manual C++ MCMC Runner Class
}
\section{Fields}{

\describe{
\item{\code{ptr}}{External pointer to C++ MCMCRunnerManual object}

\item{\code{dp_obj}}{Original Dirichlet process object}

\item{\code{distribution_type}}{Type of mixing distribution}
}}

\section{Methods}{

\describe{
\item{\code{get_auxiliary_params()}}{Get auxiliary parameters}

\item{\code{get_cluster_entropies()}}{Get cluster entropies}

\item{\code{get_cluster_likelihoods()}}{Get cluster likelihoods}

\item{\code{get_cluster_statistics()}}{Get cluster statistics}

\item{\code{get_clustering_entropy()}}{Get clustering entropy}

\item{\code{get_convergence_diagnostics()}}{Get convergence diagnostics}

\item{\code{get_iteration()}}{Get current iteration number}

\item{\code{get_log_posterior()}}{Get log posterior}

\item{\code{get_membership_matrix()}}{Get cluster membership matrix}

\item{\code{get_results()}}{Get complete results}

\item{\code{get_state()}}{Get current state of the sampler}

\item{\code{is_complete()}}{Check if all iterations are complete}

\item{\code{merge_clusters(cluster1, cluster2)}}{Merge two clusters}

\item{\code{perform_iteration()}}{Perform a complete MCMC iteration}

\item{\code{run()}}{Run all iterations}

\item{\code{sample_predictive(n_samples)}}{Sample from posterior predictive}

\item{\code{set_auxiliary_count(m)}}{Set number of auxiliary parameters}

\item{\code{set_bounds(lower, upper)}}{Set parameter bounds}

\item{\code{set_labels(labels)}}{Set cluster labels}

\item{\code{set_params(params)}}{Set cluster parameters}

\item{\code{set_temperature(temp)}}{Set temperature for annealed sampling}

\item{\code{set_update_flags(clusters = TRUE, params = TRUE, alpha = TRUE)}}{Control which parameters are updated}

\item{\code{split_cluster(cluster_id, split_prob = 0.5)}}{Split a cluster}

\item{\code{step_assignments()}}{Update cluster assignments using Algorithm 8}

\item{\code{step_concentration()}}{Update concentration parameter}

\item{\code{step_parameters()}}{Update cluster parameters}
}}

