% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_hmm_create.R
\name{DirichletHMMCreate}
\alias{DirichletHMMCreate}
\title{Create a generic Dirichlet process hidden Markov Model}
\usage{
DirichletHMMCreate(x, mdobj, alpha, beta, cpp = FALSE)
}
\arguments{
\item{x}{Data to be modelled}

\item{mdobj}{Mixing distribution object}

\item{alpha}{Alpha parameter}

\item{beta}{Beta parameter}

\item{cpp}{Logical. Use C++ implementation if TRUE, R implementation if FALSE. Default is FALSE.}
}
\description{
Create a hidden Markov model where the data is believed to be generated from the mixing object distribution.
}
