% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichletprocess.R
\docType{package}
\name{dirichletprocess}
\alias{dirichletprocess}
\title{A flexible package for fitting Bayesian non-parametric models.}
\description{
Create, fit and take posterior samples from a Dirichlet process.
}
