% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_process_beta.R
\name{DirichletProcessBeta}
\alias{DirichletProcessBeta}
\title{Create a Dirichlet Process object with Beta mixing distribution}
\usage{
DirichletProcessBeta(
  y,
  alphaPriors = c(2, 0.5),
  mhStepSize = c(0.1, 0.1),
  verbose = TRUE,
  cpp = FALSE
)
}
\arguments{
\item{y}{Data for which to be fitted}

\item{alphaPriors}{Alpha prior parameters for the DP concentration parameter}

\item{mhStepSize}{Metropolis-Hastings step size for parameter updates}

\item{verbose}{Logical indicating whether to print messages}

\item{cpp}{Logical. Use C++ implementation if TRUE, R implementation if FALSE. Default is FALSE.}
}
\value{
Dirichlet process object with Beta mixing distribution
}
\description{
Create a Dirichlet Process object with Beta mixing distribution
}
