% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_process_gaussian.R
\name{DirichletProcessGaussian}
\alias{DirichletProcessGaussian}
\title{Create a Dirichlet Mixture of Gaussians}
\usage{
DirichletProcessGaussian(y, g0Priors = c(0, 1, 1, 1), alphaPriors = c(2, 4))
}
\arguments{
\item{y}{Data}

\item{g0Priors}{Base Distribution Priors \eqn{\gamma = (\mu _0, k_0 , \alpha _0 , \beta _0)}}

\item{alphaPriors}{Alpha prior parameters. See \code{\link{UpdateAlpha}}.}
}
\value{
Dirichlet process object
}
\description{
This is the constructor function to produce a \code{dirichletprocess} object with a Gaussian mixture kernel with unknown mean and variance.
The base measure is a Normal Inverse Gamma distribution that is conjugate to the posterior distribution.
}
\details{
\eqn{G_0(\theta | \gamma) =  N \left(\mu | \mu_0, \frac{\sigma^2}{k_0} \right) \mathrm{Inv-Gamma} \left(\sigma^2 | \alpha_0, \beta_0 \right)}

We recommend scaling your data to zero mean and unit variance for quicker convergence.
}
