% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_process_mvnormal.R
\name{DirichletProcessMvnormal}
\alias{DirichletProcessMvnormal}
\title{Create a Dirichlet mixture of multivariate normal distributions.}
\usage{
DirichletProcessMvnormal(y, g0Priors, alphaPriors = c(2, 4))
}
\arguments{
\item{y}{Data}

\item{g0Priors}{Prior parameters for the base distribution.}

\item{alphaPriors}{Alpha prior parameters. See \code{\link{UpdateAlpha}}.}
}
\description{
\eqn{G_0 (\boldsymbol{\mu} , \Lambda |  \boldsymbol{\mu _0} , \kappa _0, \nu _0, T_0)  = N ( \boldsymbol{\mu} | \boldsymbol{\mu _0} , (\kappa _0 \Lambda )^{-1} ) \mathrm{Wi} _{\nu _0} (\Lambda | T_0)}
}
