% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_process_mvnormal2.R
\name{DirichletProcessMvnormal2}
\alias{DirichletProcessMvnormal2}
\title{Create a Dirichlet mixture of multivariate normal distributions with semi-conjugate prior.}
\usage{
DirichletProcessMvnormal2(y, g0Priors, alphaPriors = c(2, 4))
}
\arguments{
\item{y}{Data}

\item{g0Priors}{Prior parameters for the base distribution.}

\item{alphaPriors}{Alpha prior parameters. See \code{\link{UpdateAlpha}}.}
}
\description{
Create a Dirichlet mixture of multivariate normal distributions with semi-conjugate prior.
}
