% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_hierarchical_beta_wrappers.R,
%   R/cpp_interface.R
\name{enable_cpp_hierarchical_samplers}
\alias{enable_cpp_hierarchical_samplers}
\title{Enable C++ implementations for hierarchical samplers}
\usage{
enable_cpp_hierarchical_samplers(use_cpp = TRUE)

enable_cpp_hierarchical_samplers(use_cpp = TRUE)
}
\arguments{
\item{use_cpp}{Logical indicating whether to use C++ implementations}
}
\description{
This function enables the use of C++ implementations for the hierarchical
Beta DP sampling algorithms when available.
}
\keyword{internal}
