% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R, R/cpp_markov_wrappers.R
\name{enable_cpp_markov_samplers}
\alias{enable_cpp_markov_samplers}
\title{Enable/disable cpp markov samplers}
\usage{
enable_cpp_markov_samplers(use_cpp = TRUE)

enable_cpp_markov_samplers(use_cpp = TRUE)
}
\arguments{
\item{use_cpp}{Logical indicating whether to use C++ implementations}
}
\description{
This function enables the use of C++ implementations for the Markov
DP (HMM) sampling algorithms when available.
}
\keyword{internal}
