% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R, R/cpp_wrappers.R
\name{enable_cpp_samplers}
\alias{enable_cpp_samplers}
\title{Enable/disable cpp samplers (general)}
\usage{
enable_cpp_samplers(use_cpp = TRUE)

enable_cpp_samplers(use_cpp = TRUE)
}
\arguments{
\item{use_cpp}{Logical indicating whether to use C++ implementations}
}
\description{
This function enables the use of C++ implementations for the core
sampling algorithms when available.
}
\keyword{internal}
