% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{enable_cpp_samplers}
\alias{enable_cpp_samplers}
\title{Enable C++ implementations for specific samplers}
\usage{
enable_cpp_samplers(enable = TRUE)
}
\arguments{
\item{enable}{Logical indicating whether to enable C++ samplers}
}
\description{
Enable C++ implementations for specific samplers
}
