% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hmm.R
\name{Fit.markov}
\alias{Fit.markov}
\title{Fit a Hidden Markov Dirichlet Process Model}
\usage{
\method{Fit}{markov}(dpObj, its, updatePrior = FALSE, progressBar = FALSE, ...)
}
\arguments{
\item{dpObj}{Initialised Dirichlet Process object}

\item{its}{Number of iterations to use}

\item{updatePrior}{Logical flag, defaults to \code{FALSE}. Set whether the parameters of the base measure are updated.}

\item{progressBar}{Logical flag indicating whether to display a progress bar.}
}
\value{
A Dirichlet Process object with the fitted cluster parameters and states.
}
\description{
Fit a Hidden Markov Dirichlet Process Model
}
