% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_inverse_gamma.R
\name{GaussianMixtureCreate}
\alias{GaussianMixtureCreate}
\title{Create a Normal mixing distribution}
\usage{
GaussianMixtureCreate(priorParameters)
}
\arguments{
\item{priorParameters}{Prior parameters for the base measure.}
}
\value{
Mixing distribution object
}
\description{
See \code{\link{DirichletProcessGaussian}} for details on the base measure.
}
