% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_uniform_gamma.R, R/beta_uniform_pareto.R,
%   R/exponential_gamma.R, R/mixing_distribution_likelihood.R,
%   R/mvnormal_normal_wishart.R, R/mvnormal_semi_conjugate.R,
%   R/normal_fixed_variance.R, R/normal_inverse_gamma.R
\name{Likelihood.beta}
\alias{Likelihood.beta}
\alias{Likelihood.beta2}
\alias{Likelihood.exponential}
\alias{Likelihood}
\alias{Likelihood.mvnormal}
\alias{Likelihood.mvnormal2}
\alias{Likelihood.normalFixedVariance}
\alias{Likelihood.normal}
\title{Mixing Distribution Likelihood}
\usage{
\method{Likelihood}{beta}(mdObj, x, theta)

\method{Likelihood}{beta2}(mdObj, x, theta)

\method{Likelihood}{exponential}(mdObj, x, theta)

Likelihood(mdObj, x, theta)

\method{Likelihood}{mvnormal}(mdObj, x, theta)

\method{Likelihood}{mvnormal2}(mdObj, x, theta)

\method{Likelihood}{normalFixedVariance}(mdObj, x, theta)

\method{Likelihood}{normal}(mdObj, x, theta)
}
\arguments{
\item{mdObj}{Mixing Distribution}

\item{x}{Data}

\item{theta}{Parameters of distribution}
}
\value{
Likelihood of the data
}
\description{
Evaluate the Likelihood of some data \eqn{x} for some parameter \eqn{\theta}.
}
