% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_uniform_gamma.R, R/beta_uniform_pareto.R,
%   R/mixing_distribution.R, R/mvnormal_semi_conjugate.R
\name{MhParameterProposal.beta}
\alias{MhParameterProposal.beta}
\alias{MhParameterProposal.beta2}
\alias{MhParameterProposal}
\alias{MhParameterProposal.mvnormal2}
\title{Metropolis-Hastings Parameter Proposal}
\usage{
\method{MhParameterProposal}{beta}(mdObj, old_params)

\method{MhParameterProposal}{beta2}(mdObj, old_params)

MhParameterProposal(mdObj, old_params)

\method{MhParameterProposal}{mvnormal2}(mdObj, old_params)
}
\arguments{
\item{mdObj}{A mixing distribution object}

\item{old_params}{Current parameter values}
}
\value{
Proposed parameter values
}
\description{
Generate parameter proposals for Metropolis-Hastings sampling.
}
