% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnormal_normal_wishart.R
\name{MvnormalCreate}
\alias{MvnormalCreate}
\title{Create a multivariate normal mixing distribution}
\usage{
MvnormalCreate(priorParameters)
}
\arguments{
\item{priorParameters}{The prior parameters for the Multivariate Normal.}
}
\description{
Create a multivariate normal mixing distribution
}
