% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.dirichletprocess}
\alias{plot.dirichletprocess}
\title{Plot the Dirichlet Process}
\usage{
\method{plot}{dirichletprocess}(x, likelihood = FALSE, single = TRUE,
  y = NULL, ...)
}
\arguments{
\item{x}{Dirichlet Process Object to plot}

\item{likelihood}{Logical, indicating whether to plot the likelihood from the dpobj.}

\item{single}{Logical, indicating whether to draw the posterior from the last iteration or use the full cluster sequence.}

\item{y}{Consistency for generic.}

\item{...}{Graphical arguments}
}
\description{
Plot the Dirichlet Process
}
