% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_gamma.R,
%   R/mixing_distribution_posterior_draw.R, R/mvnormal_normal_wishart.R,
%   R/mvnormal_semi_conjugate.R, R/normal_inverse_gamma.R,
%   R/weibull_uniform_gamma.R
\name{PosteriorDraw.exponential}
\alias{PosteriorDraw.exponential}
\alias{PosteriorDraw}
\alias{PosteriorDraw.mvnormal}
\alias{PosteriorDraw.mvnormal2}
\alias{PosteriorDraw.normal}
\alias{PosteriorDraw.weibull}
\title{Draw from the posterior distribution}
\usage{
\method{PosteriorDraw}{exponential}(mdobj, x, n = 1)

PosteriorDraw(mdObj, x, n = 1, ...)

\method{PosteriorDraw}{mvnormal}(mdobj, x, n = 1)

\method{PosteriorDraw}{mvnormal2}(mdobj, x, mhDraws = 1, start_pos)

\method{PosteriorDraw}{normal}(mdobj, x, n = 1)

\method{PosteriorDraw}{weibull}(mdObj, x, n = 100, start_pos)
}
\arguments{
\item{x}{Data}

\item{n}{Number of draws}

\item{mdObj}{Mixing Distribution}

\item{...}{For a non-conjugate distribution the starting parameters. Defaults to a draw from the prior distribution.}
}
\value{
A sample from the posterior distribution
}
\description{
Draw from the posterior distribution
}
