% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_distribution_posterior_draw.R
\name{PosteriorDraw}
\alias{PosteriorDraw}
\title{Draw from the posterior distribution}
\usage{
PosteriorDraw(mdObj, x, n = 1, ...)
}
\arguments{
\item{mdObj}{Mixing Distribution}

\item{x}{Data}

\item{n}{Number of draws}

\item{...}{For a non-conjugate distribution the starting parameters. Defaults to a draw from the prior distribution.}
}
\value{
A sample from the posterior distribution
}
\description{
Draw from the posterior distribution
}
