% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_frame.R
\name{PosteriorFrame}
\alias{PosteriorFrame}
\title{Calculate the posterior mean and quantiles from a Dirichlet process object.}
\usage{
PosteriorFrame(dpobj, xgrid, ndraws = 1000, ci_size = 0.1)
}
\arguments{
\item{dpobj}{The Dirichlet process object to be drawn from.}

\item{xgrid}{The x values the posterior is to be evaluated at.}

\item{ndraws}{The number of posterior draws to take.}

\item{ci_size}{The size of the credible interval draw in terms of percentage.}
}
\value{
A dataframe consisting of the posterior mean and credible intervals.
}
\description{
Calculate the posterior mean and quantiles from a Dirichlet process object.
}
