% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_distribution_posterior_parameters.R
\name{PosteriorParameters}
\alias{PosteriorParameters}
\title{Calculate the posterior parameters for a conjugate prior.}
\usage{
PosteriorParameters(mdObj, x)
}
\arguments{
\item{mdObj}{Mixing distribution object}

\item{x}{Data}
}
\value{
Parameters of the posterior distribution
}
\description{
Calculate the posterior parameters for a conjugate prior.
}
