% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_distribution_posterior_parameters.R,
%   R/mvnormal_normal_wishart.R, R/normal_fixed_variance.R,
%   R/normal_inverse_gamma.R
\name{PosteriorParameters}
\alias{PosteriorParameters}
\alias{PosteriorParameters.mvnormal}
\alias{PosteriorParameters.normalFixedVariance}
\alias{PosteriorParameters.normal}
\title{Calculate the posterior parameters for a conjugate prior.}
\usage{
PosteriorParameters(mdObj, x)

\method{PosteriorParameters}{mvnormal}(mdObj, x)

\method{PosteriorParameters}{normalFixedVariance}(mdObj, x)

\method{PosteriorParameters}{normal}(mdObj, x)
}
\arguments{
\item{mdObj}{Mixing distribution object}

\item{x}{Data}
}
\value{
Parameters of the posterior distribution
}
\description{
Calculate the posterior parameters for a conjugate prior.
}
