% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_distribution_predictive.R
\name{Predictive}
\alias{Predictive}
\title{Calculate how well the prior predicts the data.}
\usage{
Predictive(mdObj, x)
}
\arguments{
\item{mdObj}{The distribution}

\item{x}{The data}
}
\value{
The probability of the data being from the prior.
}
\description{
Calculate how well the prior predicts the data.
}
