% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_gamma.R,
%   R/mixing_distribution_predictive.R, R/mvnormal_normal_wishart.R,
%   R/normal_fixed_variance.R, R/normal_inverse_gamma.R
\name{Predictive.exponential}
\alias{Predictive.exponential}
\alias{Predictive}
\alias{Predictive.mvnormal}
\alias{Predictive.normalFixedVariance}
\alias{Predictive.normal}
\title{Calculate how well the prior predicts the data.}
\usage{
\method{Predictive}{exponential}(mdObj, x)

Predictive(mdObj, x)

\method{Predictive}{mvnormal}(mdObj, x)

\method{Predictive}{normalFixedVariance}(mdObj, x)

\method{Predictive}{normal}(mdObj, x)
}
\arguments{
\item{mdObj}{The distribution}

\item{x}{The data}
}
\value{
The probability of the data being from the prior.
}
\description{
Calculate how well the prior predicts the data.
}
