% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{prepare_mcmc_params}
\alias{prepare_mcmc_params}
\title{Create MCMC parameters for C++}
\usage{
prepare_mcmc_params(
  dp_obj,
  its,
  updatePrior,
  n_burn = 0,
  thin = 1,
  store_history = TRUE
)
}
\arguments{
\item{dp_obj}{Dirichlet process object}

\item{its}{Number of iterations}

\item{updatePrior}{Whether to update prior parameters}

\item{n_burn}{Burn-in iterations}

\item{thin}{Thinning interval}
}
\value{
List of MCMC parameters
}
\description{
Create MCMC parameters for C++
}
\keyword{internal}
