% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{prepare_mcmc_params}
\alias{prepare_mcmc_params}
\title{Create MCMC parameters for C++}
\usage{
prepare_mcmc_params(dp_obj, its, updatePrior, n_burn = 0, thin = 1)
}
\description{
Create MCMC parameters for C++
}
\keyword{internal}
