% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_clusters.R
\name{PriorClusters}
\alias{PriorClusters}
\title{Draw prior clusters and weights from the Dirichlet process}
\usage{
PriorClusters(dpobj)
}
\arguments{
\item{dpobj}{A Dirichlet process object}
}
\value{
A list of weights and parameters of the prior distribution of the Dirichcet process
}
\description{
Draw prior clusters and weights from the Dirichlet process
}
