% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_uniform_gamma.R, R/beta_uniform_pareto.R,
%   R/mixing_distribution_prior_density.R, R/weibull_uniform_gamma.R
\name{PriorDensity.beta}
\alias{PriorDensity.beta}
\alias{PriorDensity.beta2}
\alias{PriorDensity}
\alias{PriorDensity.weibull}
\title{Calculate the prior density of a mixing distribution}
\usage{
\method{PriorDensity}{beta}(mdObj, theta)

\method{PriorDensity}{beta2}(mdObj, theta)

PriorDensity(mdObj, theta)

\method{PriorDensity}{weibull}(mdObj, theta)
}
\arguments{
\item{mdObj}{Mixing distribution}

\item{theta}{Prior parameters}
}
\description{
Calculate the prior density of a mixing distribution
}
