% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_uniform_gamma.R, R/beta_uniform_pareto.R,
%   R/exponential_gamma.R, R/mixing_distribution_prior_draw.R,
%   R/mvnormal_normal_wishart.R, R/mvnormal_semi_conjugate.R,
%   R/normal_inverse_gamma.R
\name{PriorDraw.beta}
\alias{PriorDraw.beta}
\alias{PriorDraw.beta2}
\alias{PriorDraw.exponential}
\alias{PriorDraw}
\alias{PriorDraw.mvnormal}
\alias{PriorDraw.mvnormal2}
\alias{PriorDraw.normal}
\title{Draw from the prior distribution}
\usage{
\method{PriorDraw}{beta}(mdobj, n = 1)

\method{PriorDraw}{beta2}(mdobj, n = 1)

\method{PriorDraw}{exponential}(mdobj, n)

PriorDraw(mdObj, n)

\method{PriorDraw}{mvnormal}(mdobj, n = 1)

\method{PriorDraw}{mvnormal2}(mdobj, n = 1)

\method{PriorDraw}{normal}(mdobj, n = 1)
}
\arguments{
\item{n}{Number of draws.}

\item{mdObj}{Mixing Distribution}
}
\value{
A sample from the prior distribution
}
\description{
Draw from the prior distribution
}
