% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_uniform_gamma.R, R/beta_uniform_pareto.R,
%   R/exponential_gamma.R, R/mixing_distribution_prior_draw.R,
%   R/mvnormal_normal_wishart.R, R/mvnormal_semi_conjugate.R,
%   R/normal_fixed_variance.R, R/normal_inverse_gamma.R,
%   R/weibull_uniform_gamma.R
\name{PriorDraw.beta}
\alias{PriorDraw.beta}
\alias{PriorDraw.beta2}
\alias{PriorDraw.exponential}
\alias{PriorDraw}
\alias{PriorDraw.mvnormal}
\alias{PriorDraw.mvnormal2}
\alias{PriorDraw.normalFixedVariance}
\alias{PriorDraw.normal}
\alias{PriorDraw.weibull}
\title{Draw from the prior distribution}
\usage{
\method{PriorDraw}{beta}(mdObj, n = 1)

\method{PriorDraw}{beta2}(mdObj, n = 1)

\method{PriorDraw}{exponential}(mdObj, n)

PriorDraw(mdObj, n)

\method{PriorDraw}{mvnormal}(mdObj, n = 1)

\method{PriorDraw}{mvnormal2}(mdObj, n = 1)

\method{PriorDraw}{normalFixedVariance}(mdObj, n = 1)

\method{PriorDraw}{normal}(mdObj, n = 1)

\method{PriorDraw}{weibull}(mdObj, n = 1)
}
\arguments{
\item{mdObj}{Mixing Distribution}

\item{n}{Number of draws.}
}
\value{
A sample from the prior distribution
}
\description{
Draw from the prior distribution
}
