% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_distribution_prior_draw.R
\name{PriorDraw}
\alias{PriorDraw}
\title{Draw from the prior distribution}
\usage{
PriorDraw(mdObj, n)
}
\arguments{
\item{mdObj}{Mixing Distribution}

\item{n}{Number of draws.}
}
\value{
A sample from the prior distribution
}
\description{
Draw from the prior distribution
}
