% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{PriorFunction}
\alias{PriorFunction}
\title{Generate the prior function of the Dirichlet process}
\usage{
PriorFunction(dpobj)
}
\arguments{
\item{dpobj}{A Dirichlet process object}
}
\value{
A function f(x) that represents a draw from the prior distrubtion of the Dirichlet process.

 @export
}
\description{
Generate the prior function of the Dirichlet process
}
