% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_distribution_update_prior_parameters.R
\name{PriorParametersUpdate}
\alias{PriorParametersUpdate}
\title{Update the prior parameters of a mixing distribution}
\usage{
PriorParametersUpdate(mdObj, clusterParameters, n = 1)
}
\arguments{
\item{mdObj}{Mixing Distribution Object}

\item{clusterParameters}{Current cluster parameters}

\item{n}{Number of samples}
}
\value{
mdobj New Mixing Distribution object with updated cluster parameters
}
\description{
Update the prior parameters of a mixing distribution
}
