% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_uniform_gamma.R,
%   R/mixing_distribution_update_prior_parameters.R, R/weibull_uniform_gamma.R
\name{PriorParametersUpdate.beta}
\alias{PriorParametersUpdate.beta}
\alias{PriorParametersUpdate}
\alias{PriorParametersUpdate.normal}
\alias{PriorParametersUpdate.conjugate}
\alias{PriorParametersUpdate.exponential}
\alias{PriorParametersUpdate.weibull}
\alias{PriorParametersUpdate.default}
\title{Update the prior parameters of a mixing distribution}
\usage{
\method{PriorParametersUpdate}{beta}(mdObj, clusterParameters, n = 1)

PriorParametersUpdate(mdObj, clusterParameters, n = 1)

\method{PriorParametersUpdate}{normal}(mdObj, clusterParameters, n = 1)

\method{PriorParametersUpdate}{conjugate}(mdObj, clusterParameters, n = 1)

\method{PriorParametersUpdate}{exponential}(mdObj, clusterParameters, n = 1)

\method{PriorParametersUpdate}{weibull}(mdObj, clusterParameters, n = 1)

\method{PriorParametersUpdate}{default}(mdObj, clusterParameters, n = 1)

\method{PriorParametersUpdate}{weibull}(mdObj, clusterParameters, n = 1)
}
\arguments{
\item{mdObj}{Mixing Distribution Object}

\item{clusterParameters}{Current cluster parameters}

\item{n}{Number of samples}
}
\value{
mdobj New Mixing Distribution object with updated cluster parameters
}
\description{
Update the prior parameters of a mixing distribution
}
