% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rats}
\alias{rats}
\title{Tumour incidences in rats}
\format{
\describe{
\item{y}{number of rats with a tumour}
\item{N}{total number of rats in the experiment}
}
}
\source{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari, A., & Rubin, D. B. (2013). Bayesian data analysis. CRC press.
}
\usage{
rats
}
\description{
Rat tumour data from Tarone (1982). Data from Table 5.1 of Bayesian Data Analysis
}
\keyword{datasets}
