% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_cpp_interface.R
\name{run_hierarchical_mcmc_cpp}
\alias{run_hierarchical_mcmc_cpp}
\title{Run Hierarchical Beta MCMC using C++ implementation}
\usage{
run_hierarchical_mcmc_cpp(
  dp_list,
  n_iter = 1000,
  n_burn = 100,
  thin = 1,
  update_prior = FALSE,
  progress_bar = TRUE
)
}
\arguments{
\item{dp_list}{Hierarchical DP object (not a list of DirichletProcessBeta objects)}

\item{n_iter}{Number of MCMC iterations}

\item{n_burn}{Number of burn-in iterations}

\item{thin}{Thinning parameter}

\item{update_prior}{Whether to update prior parameters}

\item{progress_bar}{Show progress bar}
}
\value{
Updated hierarchical DP object
}
\description{
Run Hierarchical Beta MCMC using C++ implementation
}
