% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{run_mcmc_cpp}
\alias{run_mcmc_cpp}
\title{Run MCMC using C++ implementation}
\usage{
run_mcmc_cpp(data, mixing_dist_params, mcmc_params)
}
\description{
Run MCMC using C++ implementation
}
\keyword{internal}
