% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_implementation.R, R/zzz.R
\name{set_use_cpp}
\alias{set_use_cpp}
\title{Set Implementation Preference}
\usage{
set_use_cpp(use_cpp = TRUE)

set_use_cpp(use_cpp = TRUE)
}
\arguments{
\item{use_cpp}{Logical indicating whether to use C++ implementations}
}
\value{
Previous setting (invisibly)
}
\description{
Switch between R and C++ implementations of core functions
}
\examples{
old_setting <- set_use_cpp(TRUE)
# Operations will now use C++ where available
set_use_cpp(old_setting) # Restore previous setting
}
