% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_alpha.R
\name{UpdateAlpha}
\alias{UpdateAlpha}
\title{Update the Dirichlet process concentration parameter.}
\usage{
UpdateAlpha(dpobj)
}
\arguments{
\item{dpobj}{Dirichlet process object.}
}
\value{
A Dirichlet process object with updated concentration parameter.
}
\description{
Using the procedure outlined in West (1992) we sample the concentration parameter of the Dirichlet process. See reference for further details.
}
\references{
West, M. (1992). Hyperparameter estimation in Dirichlet process mixture models. ISDS Discussion Paper# 92-A03: Duke University.
}
