% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_implementation.R, R/zzz.R
\name{using_cpp}
\alias{using_cpp}
\title{Check Implementation Preference}
\usage{
using_cpp()

using_cpp()
}
\value{
Logical indicating if C++ implementations are being used
}
\description{
Check whether R or C++ implementations are being used
}
