% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_hierarchical_beta_wrappers.R
\name{using_cpp_hierarchical_samplers}
\alias{using_cpp_hierarchical_samplers}
\title{Check if C++ hierarchical samplers are enabled}
\usage{
using_cpp_hierarchical_samplers()
}
\value{
Logical indicating if C++ hierarchical samplers are enabled
}
\description{
Check if C++ hierarchical samplers are enabled
}
