% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{using_cpp_hierarchical_samplers}
\alias{using_cpp_hierarchical_samplers}
\title{Check if using hierarchical C++ samplers}
\usage{
using_cpp_hierarchical_samplers()
}
\description{
Check if using hierarchical C++ samplers
}
