% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R, R/cpp_markov_wrappers.R
\name{using_cpp_markov_samplers}
\alias{using_cpp_markov_samplers}
\title{Check if cpp markov samplers are enabled}
\usage{
using_cpp_markov_samplers()

using_cpp_markov_samplers()
}
\value{
Logical indicating if C++ Markov samplers are enabled
}
\description{
Check if cpp markov samplers are enabled

Check if C++ Markov samplers are enabled
}
\keyword{internal}
