% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_wrappers.R
\name{using_cpp_samplers}
\alias{using_cpp_samplers}
\title{Check if C++ samplers are enabled}
\usage{
using_cpp_samplers()
}
\value{
Logical indicating if C++ samplers are enabled
}
\description{
Check if C++ samplers are enabled
}
