% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_censored_von_mises.R
\name{vonMisesICMixtureCreate}
\alias{vonMisesICMixtureCreate}
\title{Create a von Mises interval censored mixing distribution}
\usage{
vonMisesICMixtureCreate(priorParameters, muMargMethod = "marginal",
  n_samp = 3)
}
\arguments{
\item{priorParameters}{Prior parameters for the base measure which are, in
order, (mu_0, R_0, c).}

\item{muMargMethod}{Method for marginalization of prior mean. See
\code{\link{vonMisesMixtureCreate}}.}

\item{n_samp}{Number of Gibbs samples before we assume draws from posterior
are i.i.d. See \code{\link{vonMisesMixtureCreate}}.}
}
\value{
Mixing distribution object
}
\description{
Create a von Mises interval censored mixing distribution
}
